libqtappfw: remove deleteLater instances that aren't used
Remove several deleteLater calls that should have been delete's.

This should solve race conditions that an object calls deleteLater(), and leaves
scope before the QT main loop is called.

Outcome of the race condition is that sometimes the memory allocator reallocates it
to another message and causes a crash.

Bug-AGL: SPEC-1680
Change-Id: Ic83d8c6c204c6e3248bbe9a9def1842e1e384b2b
Signed-off-by: Matt Ranostay <matt.ranostay@konsulko.com>