libqtappfw: pbap: add import verb for contacts
To allow the switch from on every connection of updating the contact
info to per UI request of syncing.

This patchset adds the 'import' verb call. OBEX transfer are very slow
and ineffective for redundant data.

Also max_entries parameter is still left in for backwards compatiblity
but not used in the actual binding call.

Bug-AGL: SPEC-2541
Change-Id: I94e43b1fb5f6eeae3efde7d14ce7c3816cebe0ad
Signed-off-by: Matt Ranostay <matt.ranostay@konsulko.com>