fix compilation for new Qt version 5.14.1
std::hash specializations for some Qt types
are in Qt 5.14, so we dont need the custom
specialization for QString in utils.h.
Note that if this code was to be merged to
previous agl versions with pre 5.14 Qt
version, the custom specialization for
std::hash<QString> would be needed.

Bug-AGL: SPEC-3112
Signed-off-by: Raquel Medina <raquel.medina@konsulko.com>
Change-Id: I81240f4e804ff6c1046a50c486f39eec74edc057