add checks to avoid duplicates in qtappfw-bt model
Duplicate bluetooth entries have been seen on
settings app' bluetooth page, following
disconnect/connect cycles. They were caused by
multiple connected events associated to the same
device on the same adapter.
This patch avoids the duplicates in the model when
only one bluetooth adapter is present on the target,
duplicates in the model when more than one bluetooth
adapter is present is to be covered by SPEC-3294.
The current patch:
- tests the device id present on "device_changes"
events data, as this id is used to index
devices in the model.
- treats "device_changes added" events  on
an existing device in the model as if a "changed"
action was received for it.

Bug-AGL: SPEC-3424
Signed-off-by: Raquel Medina <raquel.medina@konsulko.com>
Change-Id: Ia5f273f456383880b2d855e567bdf6b41ed98352