update bt power state if no adapter
In case of no adapter, update 'powered' property to
update status fed back to clients (i.e., if there's
no adapter and the user enables bluetooth via
settings slider, this patch provides  the action to
take the slider back to OFF position, without this
fix it'd stay ON which can mislead the user's
expectations).

As SPEC-3301 has introduced changes to support the
no adapter scenario, this patch is also delivered
under this jira issue.

Bug-AGL: SPEC-3301
Signed-off-by: Raquel Medina <raquel.medina@konsulko.com>
Change-Id: I4fcc98b2a7cbf4093d39dc54ac83d5217de7e5ff