Reimplement HVAC API with VIS signals
Reimplement the HVAC interface with use of VIS signals.  There is
some impedance mismatching done between VIS signal notifications
and the API Qt signals to avoid changing the clients signficantly
for now.  If the Qt demo lifetime is extended, this may be revisited
to either switch to directly using the VehicleSignals interface or
doing a more granular mapping that works better with the VSS schema
and VIS.

Bug-AGL: SPEC-4409

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I6f1763836945600d84d6f70faea40eaa7d45ce27